resolve some reviews
