+++ pkg/kube/controller/deployflow/predicate.go
