+++ pkg/kube/types/pod/pod.go
@@ -1,8 +1,13 @@
 package pod
 
 import (
+	"context"
 	"fmt"
 	"strings"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/triton-io/triton/pkg/kube/types/workload"
 	"github.com/triton-io/triton/pkg/setting"
@@ -152,4 +157,29 @@ func (p *Pod) PodReady() bool {
 		}
 	}
 	return false
+}
+
+func SetPodReadinessGate(ns, name string, cl client.Client) error {
+	patchBytes := []byte(fmt.Sprintf(`{"status":{"conditions":[{"type":"%s", "status":"True"}]}}`, setting.PodReadinessGate))
+	return patchPodStatus(ns, name, patchBytes, cl)
+}
+
+func DeletePod(ns, name string, cl client.Client) error {
+	err := cl.Delete(context.TODO(), &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ns,
+			Name:      name,
+		},
+	})
+
+	return client.IgnoreNotFound(err)
+}
+
+func patchPodStatus(ns, name string, patchBytes []byte, cl client.Client) error {
+	return cl.Status().Patch(context.TODO(), &corev1.Pod{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: ns,
+			Name:      name,
+		},
+	}, client.RawPatch(types.StrategicMergePatchType, patchBytes))
 }
